+++ CHANGELOG.md
@@ -2,6 +2,16 @@
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [3.1.0] - 2023.10.31
+
+### Updated
+
+- Updated runtime/development dependencies
+- Updated gemspecs
+- Updated RSpec helpers namespaces
+- Updated `codeclimate`/`reek` configs
+- Updated gem version
+
 ## [3.0.9] - 2023.07.19
 
 ### Fixed
@@ -356,8 +366,8 @@ Optimized DNS (MX) validation flow. Removed needless DNS request for case when c
 
 ### Updated
 
-- Updated gem codebase, refactored `Truemail::ContextHelper`
-- Updated tests with `Truemail::DnsHelper#dns_mock_gateway`
+- Updated gem codebase, refactored `Truemail::RspecHelper::Context`
+- Updated tests with `Truemail::RspecHelper::Dns#dns_mock_gateway`
 - Updated gem development dependencies
 - Updated gem version